refactor public-link manage
