+++ src/components/Collections/CollectionShare/publicShare/manage/deviceLimit.tsx
@@ -2,15 +2,22 @@ import { Box, Typography } from '@mui/material';
 import React from 'react';
 import Select from 'react-select';
 import { DropdownStyle } from 'styles/dropdown';
+import { Collection, PublicURL, UpdatePublicURL } from 'types/collection';
 import { getDeviceLimitOptions } from 'utils/collection';
 import constants from 'utils/strings/constants';
 import { OptionWithDivider } from './selectComponents/OptionWithDivider';
+
+interface Iprops {
+    publicShareProp: PublicURL;
+    collection: Collection;
+    updatePublicShareURLHelper: (req: UpdatePublicURL) => Promise<void>;
+}
 
 export function ManageDeviceLimit({
     publicShareProp,
     collection,
     updatePublicShareURLHelper,
-}) {
+}: Iprops) {
     const updateDeviceLimit = async (newLimit: number) => {
         return updatePublicShareURLHelper({
             collectionID: collection.id,
