+++ src/components/Collections/CollectionShare/publicShare/manage/downloadAccess.tsx
@@ -1,13 +1,21 @@
 import { Box, Typography } from '@mui/material';
 import { AppContext } from 'pages/_app';
 import React, { useContext } from 'react';
+import { PublicURL, Collection, UpdatePublicURL } from 'types/collection';
 import constants from 'utils/strings/constants';
 import PublicShareSwitch from '../switch';
+
+interface Iprops {
+    publicShareProp: PublicURL;
+    collection: Collection;
+    updatePublicShareURLHelper: (req: UpdatePublicURL) => Promise<void>;
+}
+
 export function ManageDownloadAccess({
     publicShareProp,
     updatePublicShareURLHelper,
     collection,
-}) {
+}: Iprops) {
     const appContext = useContext(AppContext);
 
     const handleFileDownloadSetting = () => {
