+++ src/components/Collections/CollectionShare/publicShare/manage/index.tsx
@@ -1,12 +1,11 @@
 import { ManageLinkPassword } from './linkPassword';
 import { ManageDeviceLimit } from './deviceLimit';
 import { ManageLinkExpiry } from './linkExpiry';
-import { PublicLinkSetPassword } from '../setPassword';
 import { Stack, Typography } from '@mui/material';
 import { GalleryContext } from 'pages/gallery';
 import React, { useContext, useState } from 'react';
 import { updateShareableURL } from 'services/collectionService';
-import { UpdatePublicURL } from 'types/collection';
+import { Collection, PublicURL, UpdatePublicURL } from 'types/collection';
 import { sleep } from 'utils/common';
 import constants from 'utils/strings/constants';
 import {
@@ -15,18 +14,22 @@ import {
 } from '../../styledComponents';
 import { ManageDownloadAccess } from './downloadAccess';
 import { handleSharingErrors } from 'utils/error/ui';
+import { SetPublicShareProp } from 'types/publicCollection';
+
+interface Iprops {
+    publicShareProp: PublicURL;
+    collection: Collection;
+    setPublicShareProp: SetPublicShareProp;
+}
 
 export default function PublicShareManage({
     publicShareProp,
     collection,
     setPublicShareProp,
-}) {
+}: Iprops) {
     const galleryContext = useContext(GalleryContext);
 
-    const [changePasswordView, setChangePasswordView] = useState(false);
     const [sharableLinkError, setSharableLinkError] = useState(null);
-
-    const closeConfigurePassword = () => setChangePasswordView(false);
 
     const updatePublicShareURLHelper = async (req: UpdatePublicURL) => {
         try {
@@ -81,7 +84,6 @@ export default function PublicShareManage({
                             }
                         />
                         <ManageLinkPassword
-                            setChangePasswordView={setChangePasswordView}
                             collection={collection}
                             publicShareProp={publicShareProp}
                             updatePublicShareURLHelper={
@@ -102,14 +104,6 @@ export default function PublicShareManage({
                     )}
                 </ManageSectionOptions>
             </details>
-            <PublicLinkSetPassword
-                open={changePasswordView}
-                onClose={closeConfigurePassword}
-                collection={collection}
-                publicShareProp={publicShareProp}
-                updatePublicShareURLHelper={updatePublicShareURLHelper}
-                setChangePasswordView={setChangePasswordView}
-            />
         </>
     );
 }
