+++ src/components/Collections/CollectionShare/publicShare/manage/linkExpiry.tsx
@@ -2,16 +2,23 @@ import { Box, Typography } from '@mui/material';
 import React from 'react';
 import Select from 'react-select';
 import { linkExpiryStyle } from 'styles/linkExpiry';
+import { PublicURL, Collection, UpdatePublicURL } from 'types/collection';
 import { shareExpiryOptions } from 'utils/collection';
 import constants from 'utils/strings/constants';
 import { formatDateTime } from 'utils/time/format';
 import { OptionWithDivider } from './selectComponents/OptionWithDivider';
 
+interface Iprops {
+    publicShareProp: PublicURL;
+    collection: Collection;
+    updatePublicShareURLHelper: (req: UpdatePublicURL) => Promise<void>;
+}
+
 export function ManageLinkExpiry({
     publicShareProp,
     collection,
     updatePublicShareURLHelper,
-}) {
+}: Iprops) {
     const updateDeviceExpiry = async (optionFn) => {
         return updatePublicShareURLHelper({
             collectionID: collection.id,
