+++ src/components/Collections/CollectionShare/publicShare/manage/linkPassword/setPassword.tsx
